refactor: Point from integers to floating points

+++ src/Point.cpp
@@ -2,7 +2,7 @@
 #include <GL/glut.h>
 
 Point::Point() {}
-Point::Point(int x, int y) {
+Point::Point(float x, float y) {
   this->x = x;
   this->y = y;
 }
